requirements.txt
Flask==3.0.3
gunicorn==22.0.0
SQLAlchemy==2.0.31
PyMySQL==1.1.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
scikit-learn==1.5.1
joblib==1.4.2
